refactor(api, rendering): add `@SuppressWarnings("UnstableApiUsage")` annotations and simplify logic in registry and rendering methods

+++ src/main/java/com/peco2282/bcreborn/common/utils/BlockMiner.java
@@ -125,7 +125,7 @@ public class BlockMiner {
         hasFailed = true;
       }
     } else {
-      world.destroyBlockProgress(minerId, pos, MathUtils.clamp((int) Math.floor(energyAccepted * 10 / energyRequired), 0, 9));
+      world.destroyBlockProgress(minerId, pos, MathUtils.clamp((int) Math.floor((double) (energyAccepted * 10) / energyRequired), 0, 9));
     }
     return usedAmount;
   }
